refactor: removes getZone re-exporting
