+++ src/providers/cssDiagnosticsProvider.ts
@@ -6,7 +6,8 @@
 
 import * as vscode from 'vscode';
 import { getCSSLanguageService, DiagnosticSeverity as LsSeverity } from 'vscode-css-languageservice';
-import { buildCssDoc, getZone, getInlineStyleContext, buildInlineCssDoc } from '../utils/cssUtils';
+import { buildCssDoc, getInlineStyleContext, buildInlineCssDoc } from '../utils/cssUtils';
+import { getZone } from '../utils/zoneUtils';
 
 const cssService = getCSSLanguageService();
