+++ src/utils/cssUtils.ts
@@ -3,7 +3,6 @@
  * Imports shared zone detection from aspUtils.ts.
  */
 
-export { getZone, Zone } from './zoneUtils';
 import { TextDocument as LsTextDocument } from 'vscode-languageserver-textdocument';
 
 /**
